requirements.txt
gym==0.17.3
#numpy==1.21.0
tensorboard==2.4.0
#torch==1.7.0
#torchvision==0.8.1
pytorch-ignite==0.4.2
tensorboardX==2.1
tensorflow==2.3.1
ptan==0.7
protobuf==3.20.0
